pyproject.toml
[project]
name = "pdf_extractor"
version = "0.1.0"
requires-python = ">=3.11" 
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi>=0.115.12",
    "lxml[html-clean]>=5.3.2",
    "playwright>=1.51.0",
    "pydantic>=2.11.2",
    "requests>=2.32.3",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "packaging>=24.2",
    "filelock>=3.18.0",
    "pip>=25.0.1",
    "bleach>=6.2.0",
    "sse-starlette>=2.2.1",
    "tqdm>=4.67.1",
    "pathvalidate>=3.2.3",
    "loguru>=0.7.3",
    "markdown-it-py>=3.0.0",
    "tree-sitter-language-pack>=0.7.1",
    "typer>=0.12.5", # Added for CLI functionality
    "aiosqlite>=0.21.0",
    "python-arango==8.1.6",
    "litellm>=1.66.1",
    "redis>=5.2.1",
    "json-repair>=0.41.1",
    "rapidfuzz>=3.13.0",
    "tiktoken>=0.9.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-languages>=1.10.2",
    "docutils>=0.21.2",
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
    "pandoc>=2.4",
    "pyperclip>=1.9.0",
    "deepmerge>=2.0",
    "async-timeout>=5.0.1",
    "tenacity>=9.1.2",
    "jupyter>=1.1.1",
    "jupyter-contrib-nbextensions>=0.7.0",
    "tabulate>=0.9.0",
    "robotsparser>=0.0.12",
    "readability>=0.3.2",
    "readability-lxml>=0.8.1",
    "spacy>=3.8.5",
    "ftfy>=6.3.1",
    "markdownify>=1.1.0",
    "camelot-py[base]>=1.0.0",
    "marker>=2.1.3",
    "pdf2image>=1.17.0",
    "gunicorn>=21.2.0", # Added for production server
    "pillow>=11.2.1",
    "torch>=2.2.2", # For CUDA 12.8 compatibility
    "torchvision>=0.17.2", # Matching version for torch 2.2.2
    "torchaudio>=2.2.2", # Matching version for torch 2.2.2
    "llm>=0.24.2",
    "paramiko>=3.5.1",
    "modelcontextprotocol>=0.1.0",
    "mcp[cli]>=1.6.0",
    "fastapi-mcp>=0.3.2",
    "pytrie>=0.4.0",
    "pymupdf>=1.25.5",
    "html2text>=2025.4.15",
    "reportlab>=4.4.0",
]

[project.optional-dependencies]
dev = ["ruff"]
test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "respx>=0.22.0",
    "docker>=7.1.0",
    "pytest-dependency>=0.6.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.1"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "dependency: mark test to run after specified dependencies",
]
